requirements.txt
jsonschema==4.23.0
flask==3.1.0
groq==0.25.0
requests==2.32.3
python-dotenv==1.2.2
